% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckInput.R
\name{CheckInput}
\alias{CheckInput}
\title{Check if the input is valid or not}
\usage{
CheckInput(X, f, p)
}
\arguments{
\item{X}{A m*p matrix which each row represents one unique model with the elements either 0 or 1.}

\item{f}{A vector with m elements contain each model's frequency in X.}

\item{p}{The number of variate in the model}
}
\value{
The standardized matrix
}
\description{
Input a valid matrix
}
