% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSM_plot.R
\name{DSM_plot}
\alias{DSM_plot}
\title{DSM_plot
plot the naive visualization of the distribution of selected model}
\usage{
DSM_plot(
  X,
  f,
  p,
  Anchor.model = NULL,
  circlesize = NULL,
  linewidth = NULL,
  fontsize = NULL
)
}
\arguments{
\item{X}{A m*p matrix which contains m different p-dimensional models. All the elements are either 0 or 1.}

\item{f}{A vector with m elements which represent each model's frequency in X.}

\item{p}{The number of variate in the model}

\item{Anchor.model}{A vector containing p elements with either 1 or 0 value and must be found in X. Default is the model with the highest frequency.}

\item{circlesize}{customize the size of the circle in the plot, default is 10.}

\item{linewidth}{Customize the width of the line in the plot, default is 1.}

\item{fontsize}{Customize the size of the font in the circles, default is 1.5.}
}
\value{
A summarized information of the grouped models.
}
\description{
DSM_plot
plot the naive visualization of the distribution of selected model
}
\examples{
data(exampleX)
X=exampleX
data(examplef)
f=examplef
p=8
DSM_example1 = DSM_plot(X,f,p)
}
