% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_H_info.R
\name{Groupinfo}
\alias{Groupinfo}
\title{Group the models according to their Hamming distance and Model complexity to the anchor model}
\usage{
Groupinfo(X, f, p, Anchor.model = NULL)
}
\arguments{
\item{X}{A m*p matrix which contains m different p-dimensional models. All the elements are either 0 or 1.}

\item{f}{A vector with m elements which represent each model's frequency in X.}

\item{p}{The number of variate in the model}

\item{Anchor.model}{A vector containing p elements with either 1 or 0 value and must be found in X. Default is the model with the highest frequency.}
}
\value{
A summarized information of the grouped models.
}
\description{
Group the given models
}
