% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter_heat_info.R
\name{VDSM_scatter_heat}
\alias{VDSM_scatter_heat}
\title{VDSM-Scatter-heatmap-info}
\usage{
VDSM_scatter_heat(X, f, p, Anchor.estimate, Anchor.model = NULL)
}
\arguments{
\item{X}{A m*p matrix which contains m different p-dimensional models. All the elements are either 0 or 1.}

\item{f}{A vector with m elements which represent each model's frequency in X.}

\item{p}{The number of variate in the model}

\item{Anchor.estimate}{An estimation for the anchor model}

\item{Anchor.model}{A vector containing p elements with either 1 or 0 value and must be found in X. Default is the model with the highest frequency.}
}
\value{
A list of information which helps to plot VDSM-Scatter-heatmap.
}
\description{
Report VDSM-Scatter-heatmap-infomation
}
